Add Dynatrace Monitoring as Code v1.0
The first OpenSource release of Dyntrace Monitoring as Code, a tool
that allows configuring Dynatrace in an "infrastructure as code"-like way.

This intial commit squashes all internal development done before
opensourcing Dynatrace Monitoring as Code.

As such it collects a team effort of core contributors, as well as
many others that helped develop, test and use the tool interally.

Co-authored-by: didiladi <dieter.ladenhauf@dynatrace.com>
Co-authored-by: tatomir146 <sinisa.tatomir@dynatrace.com>
Co-authored-by: CruzanCaramele <hamza.yahaya@dynatrace.com>
Co-authored-by: goberchtold <berchtold@student.tugraz.at>
Co-authored-by: patrickpichler <patrick.pichler@dynatrace.com>