Temporarily run integration tests just on linux
This commit changes the workflow which test executed when building
integration tests of PRs coming from forks (Run Integration Tests
for Forks). It temporarily disables builds to run on MacOS and Windows.

The reason for that is #30, builds are currently failing on windows.
Furthermore, there seems to be a problem with concurrent integration
test execution.

We should revert this change again anlongside the fix for #30.