Log a warning if project name clashes with API name (#58)
This commit introduces a log message if the project folder name clashes
with an API folder of the same name, but different lowercase/uppercase
spelling. Additionally, the readme was changed, to indicate that API
folders need to be defined case-sensitive.

Example of a log line:
```
2020-11-30 10:43:07 WARN  Project Application in folder \
skip-deployment-project clashes with API name application. \
Consider using a different name for your project.
```

fixes #58