Encapsulate REST code in DynatraceClient (#74)
This commit introduces a new client (`DynatraceClient`) for performing
basic CRUD operations on any Dynatrace API supported by monaco.

It encapsulates the configuration-specific inconsistencies of certain
APIs in one place to provide a common interface to work with. After
all: A user of DynatraceClient shouldn't care about the implementation
details of each individual Dynatrace API.

Its design is intentionally not dependant on the Config and Environment
interfaces included in monaco. This makes sure, that DynatraceClient
can be used as a base for future tooling, which relies on a
standardized way to access Dynatrace APIs.

The new client implementation opens the door for future improvements,
like concurrent web request execution and the implementation of
various caching strategies.