Introduce new experimental cli design
The new cli design features subcommands instead of solely relying on
flags. In order to use the new CLI, on have to set the `NEW_CLI` env
variable to 1. Without this variable set, the old design is active.

The plan is to deprecate the old cli with version 1.5 and remove it with
version 2.0.

This will fix #45.