Fix #60: Allow path separators in variables (#83)
Changing the conditions that need to be met in order
to make the unmarshal function to replace slash or backslash with
platform-dependant path separator. Otherwise, assume that it is not
a path and therefore normalizing is not needed (e.g. in case of a
user agent string).

Only normalize slashes and backslashes in JSON references that appear in
the YAML config section. Do not touch them if they appear in another
YAML section.