Fix #77: Enhance log messages for syntax errors detected in JSONs (#88)
This commit enhances the log messages displayed to the user if a json
contained a syntax error. Previously, we just displayed the message of
the original cause to the user. This resulted in a error message like
this:

```
2020-12-01 14:36:10 ERROR Validation of sandbox failed with error \
  operation.json is not a valid json! Error: invalid character '"' \
  after object key:value pair
```

This commit introduces a second log which indicates where in the
file the syntax error is. Besides the line and character numbers,
we also display now line (+previous line), which contain the syntax
error.

E.g. if the user forgets to put a comma after a key:value pair:

```
2020-12-07 19:52:12 ERROR File did not contain valid json:
 --> operation.json:7:4
   |    "no": "comma"
 7 |    "boolean": true
   |    ^^^
   - Cause: invalid character '"' after object key:value pair
```

E.g. if the user forgets the closing curly brace for an object:

```
2020-12-07 19:52:12 ERROR File did not contain valid json:
 --> test.json:6:2
   |    "foo": "bar"
 6 |  ]
   |  ^^^
   - Cause: invalid character ']' after object key:value pair
```

This log line format was heavily inspired by the Rust compiler.