Cope with faulty configs (name or id is null)
It can happen that you have a faulty config deployed in your
environment, which has no id or name set. These special cases were not
handled correctly before. Instead, `monaco` crashed when getting the
list of configs. This commit introduces a graceful handling of these
configs t monaco.

If the id is null -> Return an error
If the name is null -> Fill in the id as the name (as done by the
extensions API)