Add response log
Since it is sometimes also useful to see the response of certain
requests, there is now the possibility to specify a response log file
via the MONACO_RESPONSE_LOG env variable. It basically works exactly as
the request log.

The log output format has changed a bit and is now bsically the raw dump
of the outgoing requests/response.

This fixes #228