Do not unmarshall json before sending it to api
Unmarshalling the json before sending it to the dyantrace api is causing
issues. For example it will cause numbers to be transformed from int
values to float values. In order to not introduce more issues, we no
longer unmarshall the resulting json to a map and instead directly send
it as is to the dynatrace api.