fix: Paging in monaco was not working
When having a lot of SLOs defined on the environment, monaco just
grabs the first 10 configs when performing the GET (all) request
to determine if a config is already defined. As a result, monaco
performs a POST instead of a PUT for the SLO config.

This commit fixes this behavior and fixes teh paging functionality.

Additionally, it extends the `all_configs_integration` test with a
check which applies the same configs again, to enforce a PUT
(update) of the just created configs.

fixes #276