feat: Check Version before Extension upload
Add a step to extension_upload that looks for a currently deployed,
for the given extension name.

If the versions are equal, or the config as code version is lower,
skip the deployment, otherwise deploy the extension as before.

This removes the very user-unfriendly attempt to upload already
present versions, resulting in an incorrect "Error" log, and also
logs actual information about the already deployed version of the
extension.

fix: #125