feat: Add own logger implementation
This is to get rid of the usage of the currently used logger which is no
longer actively maintained.

The goal was to make a very simplistic logger which supports only what
we need, so the only exported functions are:
* A setup function, which should only be needed to be called once
* The logging functions for actual logging. There's no need for an
  instance of a logger.
And that's it. There's still the concept of verbose logging, which
enables log.Debug prints for stdout. Log files always contain all logs.