feat(test): Add the first integration test for config v2
This commit introduces the first integration test for config 2.0. The
first test is the `AllConfigs` integration test, which tries to apply
all configs supported by monaco.

Furthermore, this commit extends the manifest model to support passing
the environment URL as environment var:

```
projects:
- name: project
- name: only-post
environments:
- group: default
  entries:
  - name: environment1
    url:
      type: environment
      value: URL_ENVIRONMENT_1
    token:
      name: TOKEN_ENVIRONMENT_1
```

This feature is needed to not leak environment URLs in gour GitHub
workflow.

fixes #387