feat(api): calculated metrics for mobile, web, synthetic, service
This commit builds upon the work of @Dynatrace-Adam-Gardner and adds the
following new APIs to monaco:

  * Calculate metrics for web applications
  * Calculate metrics for mobile applications
  * Calculate metrics for synthetic tests
  * Calculate metrics for services

Additionally, it adds the new APIs to the table of token permissions in
the documentation and adds the integration test configs for the new
APIs.

For the calculated metrics of mobile apps, I temporarily defined a
mobile application (manually created) in the environment used for the
integration test. This needs to be revised, when fixing #275.