feat(discussion): Start a discussion about config v2 models
This commit intends to start a discussion about the model in config v2.
It contains the following changes:

  * Add a version in the top level definition of configs: This helps us
to make changes to the config after the initial release. By not doing
that, we will have a hard time with avoiding backward compatibility
breaking changes. By introducing a version number, we can support
different versions of the config at the same point in time. This helps
us with adding new features and deprecating old ones.

```
version: v2
configs:
- ...
```

  * Remove the config id. The need to specify a config id  seems to me
like an unnecessary hurdle which only makes things more complicated. We
risk that users mistage this for the MeId or ConfigIdentifier coming
back from the Dynatrace APIs. The question here is, if this id is really
necessary?

```
version: v2
configs:
  config:
    name: Federation Availability
```

  * Remove the need of using `interface{}` as the type for
`configDefinition`'s `Name` and `Parameters` properties. Using
`interface{}` as type has the main disadvantage that the config has no
static model anymore - any syntax is valid (as long as its valid yaml).
The user has no possibility to check if the config matches the correct
syntax. That's why this commit aims to split the name and the reference
intwo different properties: `Name` (type: string) and `NmaeFromReference`
(type: `ConfigParameterType`). One of the two properties needs to be
set:

```
version: v2
configs:
- id: availability
  config:
    name: Federation Availability
    template: availabilty.json
    skip: false
```

```
vrsion: v2
configs:
- id: availability
  config:
    nameFromRference:
        api: management-zone
        config: zone
        property: name
        type: reference
    template: availabilty.json
    skip: false
```

This change makes it possible to add linting / model cheing in IDEs
later on.