feat(workflow): Add workflow for generating bill-of-materials
This commit adds a workflow for generating the bill of materials for
`monaco` and its dependencies (including the documentation). The
workflow needs to be started manually for now. At a later point in time,
we can run it for every release. But for now: let's keep it simple.

The approach is "borrowed" from `keptn/keptn` :)

fixes #466