refactor(rest): Build full URL in upsert method rather than caller
Move the building of the full URL from api and environment URL into the upsert method.

This is possible as the api is known to the config upsert, unlike the extension upsert method.

Passing the environment URL into upsertDynatraceObject rather than the full URL, enables using the new
version client inside config_upload.
This keeps the encapsulation of any specific handling needs for an API inside the upsert method.

The version check will be needed to match a change to the app-detection-rule API.