feat(rest): Replace app-detection-rule names with UUID on Dynatrace environment versions >= 1.236
Starting in version 1.236.0 the app-detection-rule API no longer uses a name property.
If supplied with a 'name' that is a UUID, it will use this identifier,
any other string will result in a generated UUID identifier.
Dynatrace will return the same UUID as id and name when configuration is queried.

For backwards compatibility and to match this in monaco, migrated existing rules as well as rules configured via
monaco use a generated UUID v3 derived from the name property.
For monaco this replacement is made transparent by only using the UUID at the interface to Dynatrace - when checking
for existing configuration and in the uploaded configuration playload - while otherwise using the unique objectName
as for any other configuration.