refactor: Simplify config upload error handling
Move the logic of whether deployment errors of a single config are returned or just appended for later,
out into the main execute method.

This removes the need for executeConfig to make this decision and changes it to just return an error or not.
With that the side-effect of executeConfig modifying the errors slice is removed.

The previous if/else of handling the error is also simplified so that the appending of the new error,
which has to happen in either case happens once, before returning if not explicitly continuing on errors.