feat: Handle Dynatrace Configurations with non-unique names
Some Dynatrace APIs/Configurations do not fulfill the monaco assumption of having a unique name. 
For these configurations it is possible or even required to create several configurations with the same name. 

This causes problems for monaco during deployment - being unable to update the correct configuration - and on download - being unable to download all configurations. 

This adds logic of addressing such non-unique-name configurations by their UUID - either from Dynatrace if downloaded, or generated from monaco if first uploaded. 

To stay backward compatible this handling of non-unique APIs is introduced in -v2 configuration folders, which existing configuration will keep functioning in the existing - but wrong if more than one configuration of a name exists - way.

## Commits: 
* feat: Flag APIs that don't work with unique names

* feat: Add function to test and generate UUID

* feat: Allow force updates by entity id

* feat: Allow generating project unique uuids

* feat: Use id json if name not unique

* feat: Read/update config instead of overwriting

* feat: Add deployment handler. add tests

* fix: Skip deleting non-uniquely named entities

* docs: Add info about non-unique name APIs

* fix: Fix several golangci-lint complaints

* feat: Resolve several review sugestions

* feat: Remove uuid generator from project interface

* feat: Mitigate entity duplication with v2 configs

* test: Fix non-nullable attributes

* feat: Flag APIs that don't work with unique names

* feat: Add function to test and generate UUID

* feat: Allow force updates by entity id

* feat: Allow generating project unique uuids

* feat: Use id json if name not unique

* feat: Read/update config instead of overwriting

* feat: Add deployment handler. add tests

* fix: Skip deleting non-uniquely named entities

* docs: Add info about non-unique name APIs

* fix: Fix several golangci-lint complaints

* feat: Resolve several review sugestions

* feat: Remove uuid generator from project interface

* feat: Mitigate entity duplication with v2 configs

* test: Fix non-nullable attributes

* feat: Explicitly test UUID format

* docs: Add deprecation infos

* docs: Fix broken link

* docs: Update documentation/docs/Guides/add_new_api.md

Co-authored-by: Nico Riedmann <UnseenWizzard@users.noreply.github.com>

* docs: Update documentation/docs/Guides/add_new_api.md

Co-authored-by: Nico Riedmann <UnseenWizzard@users.noreply.github.com>

* feat: Fix comment; Remove obsolete func parameters

* test: Add TestIsDeprecatedApi

* feat: Remove NewIdValue function

* feat: Undo deploy refactoring;

* feat: Reduce calculated cyclomatic complexity

* refactor: Simplify config upload error handling

Move the logic of whether deployment errors of a single config are returned or just appended for later,
out into the main execute method.

This removes the need for executeConfig to make this decision and changes it to just return an error or not.
With that the side-effect of executeConfig modifying the errors slice is removed.

The previous if/else of handling the error is also simplified so that the appending of the new error,
which has to happen in either case happens once, before returning if not explicitly continuing on errors.

Fixes #642, fixes #377, fixes #254, fixes #306

Co-authored-by: Tobi Gremmer <tobias.gremmer@dynatrace.com>
Co-authored-by: tobigremmer-dt <83594825+tobigremmer-dt@users.noreply.github.com>