fix: Ensure v1 integration tests use abs filepaths
With the introduction of always working with absolute paths,
the tests previously downloading to or overwriting files on
a relative path in the in-memory-fs fail.

Apparently the in memory layer is not aware of the correct absolute path,
and will only find the downloaded resources at the path they where written to.

To fix this, all tests use absolute path to get around the
limitiations of the in-memory FS implemtation.