fix(ci): Append actual random number to integration test names
To ensure there's no collisions between integration test configs, a timestamp
and random number are appended to all names.

Before this fix the pseudo random generator was not seeded and would produce
colliding numbers still not fixing the issues.

Seed the generator with the current timestamp down to nano seconds to ensure
we get actually non colliding test configuration names.
Additionally increase the range of numbers from 100 to 10000.