refactor: Make it possible to pass filesystem into integration test runs
This is needed if a memory overlay filesytem (CopyOnWrite on top of Os Filesystem as used in test utils),
is meant to already contain modified files before being passed into the deployment logic used in integration
testing (first rewrite files to ensure unique names, then deploy, then cleanup).

This is e.g. needed for testing conversion of a v1 project into a v2 project that can be deployed.