refactor: Remove pkg/errors dependency
pkg/errors is only ever used to create new errors, which
can be fully replaced with core fmt.Errorf functionality.

One use of wrapped errors that was only present in testing,
can also be accomplished by Errorf and %w since Go 1.13 -
which this was not possible when monaco development started,
this being a test, there was likely never a need for use of
the now archived pkg/errors for a single package.