fix: Workaround synthetic API failing if Application ID can not (yet) be found
The synthetic API can currently fail with a 500 internal server error,
if an unknown Application is referenced.

This can also happen if a newly created Application is referenced,
but can not yet be found (some delay to propagate new IDs).

To workaround this, extend the known error retry for synthetics, to
also retry on 500 rather than the currently expcted 'Unknown Application'  reply.