fix: Continue over individual delete failures
Also improve some log messages as part of aggregated error logging.

This change brings delete in line with the general idea to continue
as far as possible also applied to deployments. In this case, failure
to delete a single defined object, does not mean that the rest can not
be deleted.

ref: #720