fix(ci): Clear json template cache explicitly before converion E2E test
The test was previously failing because of a missmatch between the contents of
v1 and already converted v2 test resources which lie at the same project filepath.

This could lead to the conversion writing config jsons referencing variables, that
where not part of the v1 config yamls, leading to failing tests when trying to fill
these templates.

As there is open discussion if the template caching could just be kicked (premature
optimization), this is a quickfix to get the full CI back. Full rework of testing
methods or change to the template registry droping the cache to follow.