refactor: Drop template cache
Caching only is helpful in a few cases (if a json is actually referenced more than once)
and brings only a small memory improvements as two maps - template struct and actual go
templates - are head in memory even though they moght not be needed anymore.

Furthermore the caching based on paths from manifest root actually leads to issues of
fetching wrong templates if files with different contents share an ID - this mostly
(only) caused issues for e2e tests, but emphasises that the complexity outweighs the
benefit of optimizing.

Removing the caches offloads parsing and rendering of the actual templates to when
they are needed as prot of the Render emthod.