refactor: Simplify template allocation/adressing
Previously we used new to allocate and get a reference to a fileBasedTemplate variable,
which we immediately filled with the content of a struct.

This is technically equivalent to just addressing the struct (&) in this case.

To make code easier to read and reason about, this simplifies the allocation by
first just creating the struct - and keeping the variable in the method scope as
a variable rather than a pointer - and then adressing it only when it's actually needed
on the return call.

For details on new vs adressing for structs see: https://stackoverflow.com/a/69785168
For further team internal discussion on this see: https://dynatrace.slack.com/archives/C03QE4VGD43/p1663163543504359