fix: Escape all characters for JSON templating
It was found that it's possible to define strings in yaml configs that contain
double quotes, which will result in invalid JSONs once templated.

Instead of adding logic to escape double quotes (" -> \") in addition to the
existing logic to escape linebreaks, code was simplified to use json marshalling,
to insure that properties can savely be placed in JSON payloads when templating.