test: Extend special character config E2E test
The newly improved logic to escape config properties usings json marshalling
before rendering them into a JSON template, allows the existing E2E test to
be re-activated.

Additionally extend the test configuration with cases for newlines and
double quotes that need to be escaped for templating and deployment to work.

fixes: #121
ref: #741