fix: Include already read response body in optional response log
When response logging is active, the whole repsonse object is dumped
using methods from the httputil package. While these methods take care
to retain the content of the response body after reading, our application
code happening before does not.

With that by the time the response logging happens the body is read and
the reader closed.

To keep writing simple and not add logic retaining the body just for
optionally logging it again, the string body is simply passed to the
response logging code to write directly.