feat(download): Create full project including manifest when downloading
Instead of directly writing the downloaded project folder,
downloads are now written into an output folder which contains
projects as well as a manifest.yaml.

If a manifest.yaml already exists in the output folder, a
timestamp is appended and a warning logged to inform the user
to manually select the manifest they want to keep.