feat(download): Add DynatraceClient decorator that limits parallel access
As we limited the amount of parallel access, it was possible to run into a deadlock. E.g. 1 is allowed, and we block it for starting one api-gathering, no download would have happened.

With this approach, the actual DynatraceClient is now limited, so as long as the Dynatrace Client returns something, no dead lock is possible.


Signed-off-by: David Laubreiter <david.laubreiter@dynatrace.com>