fix(rest): Do not handle 404 response as delete error
If the API reports a given object as non-existing we can assume
successful 'deletion' for the sake of any code wanting a config to be gone.

Thus this is handled the same as trying to delete an object by name that
can not be found.

Both of these cases should not trigger error handling of callers assuming
that a delete error means the object still exists .