test(download): Fix flaky limiter test
The test was flaky because the channel was defer-cleared inside a goroutine after the callback has been triggered, and the check that the channel is empty might happen before the channel has been cleared.
Now the test only tests whether all callbacks have been executed, not that the channel is empty.


Signed-off-by: David Laubreiter <david.laubreiter@dynatrace.com>