refactor(test-resource): Use enviornment var without quotes in conversion source
It is allowed in v1 (or yaml) in general to not put quotes around string values,
this is relatively frequently used in v1 environment variable references, which
will be replaced with a full string on templating the file.

However this causes issues when loading without templating for conversion.
To insure subsequent handling of this case stays tested, the test resource is
adapted.