feat: Allow loading project files with or without templating
To enable loading projects without templating of YAMLs (needed to convert env var references),
an unmarshal function type is added and such a function is passed through project loading to
control wether tempating happens on project loading or not.

As conversion has fewer requirements than loading for deployment a dedicated method to load
projects for conversion is introduced, which does not template YAMLs on load, and skips project
sorting and the option to filter for specific projects and their dependencies.

Additionally to catch an edge case in how monaco configuration is written, any non-quoted
references in a yaml are surrounded by double-quotes if loaded without templating.
(templating does produce valid yaml in this case, as references are replaced with environment
variable values).