fix(config): Do not require Name when writing skipped configs
It is a quite frequent usecase in real-world configurations to omit the required
'name' property when a config is skipped.
Before this writing such configurations - after convert - would however fail as
writing checks for the name to be present.
If no name is present for a skipped config a default name is written instead.