docs: Release 1.8.6
Release 1.8.6 fixes an issue of monaco not ignoring hidden sub-folders - like it's own .logs/ - in some cases.

Up to now, hidden directories were ignored only if they were in the 'top' level of a loaded project.
If a hidden directory was inside a sub-folder below that, it was not.

With this fix:
1) hidden directories can no longer be projects
2) having hidden directories inside project folders works now

* fix: Project loading should ignore hidden directories

* chore(deps): bump github.com/urfave/cli/v2 from 2.20.2 to 2.23.0
* chore(deps): bump github.com/urfave/cli/v2 from 2.17.1 to 2.20.2
* chore(deps): bump github.com/urfave/cli/v2 from 2.16.3 to 2.17.1
* chore(deps): bump golangci/golangci-lint-action from 3.2.0 to 3.3.0
* chore(deps): bump actions/checkout from 3.0.2 to 3.1.0