fix: Mark app-detection-rule as actually having unique names
As discovered as part of RFA CA-1876 the migration of app-detection-rules to Settings 2.0
actually changes the APIs behaviour to having unique names again - both for creation via the UI
(as names are now set to the rules UUID rather than the application name) and on the API,
which now actually completely disregards any UUID sent on PUT, breaking the workaround for
handling non-unique-apis, but ensuring that a given rule for a specific Application will get the
same UUID.

With this the API fully behaves as a normal API from monaco's viewpoint again.

As this migration happened with Dynatrace 1.252 we're assuming that using v2 configuration
will not encounter the previous issue of UI created rules having overlapping names.