feat(delete): Allow deletion of non-unique-name-apis
This change makes it possible to delete entities which have non-unique names.
If we have a non-unique name api, we check whether the name occurs only
zero: Check if the name is actually an id. If it is, delete by id.
once: Just delete it by name
multiple: Return an error, must not be deleted.


Signed-off-by: David Laubreiter <david.laubreiter@dynatrace.com>