refactor(ci): Extract artifactory push into function
Additionally we now securely pass credentials to the shell script call.

Instead of using groovy string interpolation, which passes the credentials values to the
underlying processes and would make them visible on OS level, pass every required input as
environment variables and only resolve them on an OS level instead of on the groovy string.