fix(download): Don't double-escape ids
Fetching an entity by id (ReadId) has two possible entry-points:
1) Download -> ReadById
2) ReadByName -> ReadById

The second one works fine, as the id is passed directly in ReadById without encoding. The first one does encode it before passing it to ReadById, thus double encoding it.

This commit fixes the double encoding, so that it is now only encoded in ReadById.

Signed-off-by: David Laubreiter <3740651+Laubi@users.noreply.github.com>