test: Fix download circular dependency test
The test was previously only failing if an unexpected error occured, but not
validating in the expected error case. Thus it also kept passing after the
error message was modified.

This refactors it back to returning an error, to remove the harder to reason
checks for (non)empty string returns marking an issue and changes the test
to correctly assert in both cases where an error is expected and none should
occur.