feat: Implement Settings Client
A new SettingsClient is introduced as an abstraction layer for CRUD operations on the Dynatrace Settings API.
The methods are not added to the existing DynatraceClient, as the behavior of the APIs differ in multiple ways, most importantly on the payload level. The DynatraceClient sends the payload as 'is-as', the SettingsClient needs to wrap it in additional information. This additional information is also not present in the DynatraceClient.
It would have been possible to add new methods handling this to the DynatraceClient, but its own interface segregates the responsibilities nicely, as more methods will be added in the future.