refactor: Restructure clients in rest package
Renamed DynatraceClient to ConfigClient and introduced a new DynatraceClient that combines the SettingsClient and ConfigClient.
This allows to easily integration in existing code but still splits the interfaces for their concerns.
Other implementations of the DynatraceClient only need to add the new methods instead of adding a new implementation for each. Updated those implementations already.