fix: Correctly handle pagination when other query parameters are set
The addition of an extra query parameter for getting all SLO configurations, lead
to discovering that we do not correctly handle pagination if any other query params
are set.

Also extend the previously added to tests to fully validate the behaviour of the method
getting all configs from the API, encluding what parameters are set based on defined
server returns and whether we expect an error or success.