feat: Introduce retry mechanism for GET requests
Testing the SLO APIs pagination it was found that it's possible to run
into the same timing issues as Upsert when listing existing configs.
To handle cases where one reply returns a nextPage key/number of configs
that does not match the one of the next reply, GET is now retried.