fix: Using PreRun func instead of custom wrapper funcation
The fix/hack of using a custom wrapper function around the cobra command did not work well with assertions in integration tests since internally we are using Run instead of RunE. This caused difficulties in asserting that a command raised an error. With this fix SilenceUsage is set in a separate PreRun hook.