refactor(rest): Generalize and extract pagination handling
Pagination handling was previously only used in config_client and applied
rules specifically to SLO APIs when setting next page key query parameters.

Extract handling into it's own file for re-use between config and settings
clients.

Generalize pagination query parameter logic to work on all api/v2 APIs -
which follow a pattern of requiring a nextPageKey as the ONLY query parameter,
as it already encodes previous query parameters.
Theoretically the setting of nextPageKey as a query parameter will also work
on APIs following the (less standardized) old model of still requiring ALL
previous parameters when querying the next page - but for those APIs extracting
the nextPageKey from response body does not work, as they usually send the key
as a Header. (This is a non-issue as mostly environment APIs not covered by
monitoring as code use this pagination model).