refactor(rest): Encapsulate pagination/nextPageKey parsing in Respobse object
Previoulily a caller needed to be aware of the method offered in pagination.go,
and unmarshal response bodies to map[string]interface{} explicitly to get the
next page key if it exists.

This moves the unmarshalling and parsing into the shared code returning our custom
Response struct, which now contains the nextPageKey (possibly empty) and is extended
with a simple bool method to check if the response was paginated.