feat: Allow scope to be a configurable as a parameter
This change introduces the possibility that a config's settings-scope may be either a value or reference. This allows the use-case that we create settings for an entity-scope that has been created in another config.
Additionally, this change introduces the possibility to reference the scope of another config to chain scopes.

How this is implemented:
* The persistence is changed to load parameters instead of plain strings for the scope. We reuse the existing serializer & deserializer for this. This also already gives us the shorthand syntax.
* Within the in-memory representation, the scope is moved from the 'type' to a parameter. This has several advantages: We do not need to care about topology sort, we can access it like the name, and the resolution also already works out of the box.