ci: Checkout PR state if labeled to run e2e tests
Note: while this is not 100% safe (and it is discouraged to checkout PRs when running
in the context of the pr target repo), we need access to secrets to run E2E tests.

This is simpler but functionally the same to the previous solution using two actions,
which would also check out the PR and run in our repo context after a label was found.

We achieve some security by manually setting the label only after having reviewed an
external contribution, and by limiting the actions permissions to read only.

Further reading on actions security: https://securitylab.github.com/research/github-actions-preventing-pwn-requests/