fix(deploy): Ensure the actual manifest path can still be logged if getting absolute path produces error
Previously we possibly logged an empty string, as filepath.Abs returns one in case of
error and we directly overwrote the content of the manifest path variable.

Output of filepath.Abs is now assigned to a dedicated variable and the
non-absolute path is retained and logged in case of error.