ci: Mark E2E checkout of PR branch as ignored by semgrep
If a PR was reviewed and deemed safe to run in the context of our repo and it's
secrets, we label it to trigger E2E tests.
In that case this Action is triggered in pull_request_target context and checks
out the HEAD of the PR branch.
This is a semi-secure manually reviewed way to ensure we only run code we're
fine accessing our secrets.

Thus we are ok with the security risk that semgrep is flagging
for this action.