chore: refactored settings client and UpsertSettings (#23)
* chore: refactored settings client and UpsertSettings

"settingsClient::ListKnownSettings" is removed in favor of a single method "ListSettings" that can be used for all current use cases. Further, UpsertSettings is now handling doing the check for currently known Settings. Hence, the newly introduced type "KnownSettings" is not used anymore and was removed...

* chore: introduced discardValue field in ListSettings options

this supports the use case where we do not actually care about the "value" field of the fetched settings objects

* chore: added additional test covering filtering of returned settings objects

* chore: Use DiscardValue property when listing settings inside UpsertSettings

* chore: remove outdated comment