feat: Introducing `--specific-settings | -s` to CLI (#38)
* feat: added --specific-settings to cli and split Download function into Download and DownloadAll functions

* feat: make specific-api and specific-settings self contained

This commit contains changes to use `--specific-settings` in a way that no api's are downloaded and vice versa. It's also possible to use both at the same time.

The api map is now not part of the options anymore but is injected into the respective download function. The option structs only holds specific api names and specific setting schemas that should be downloaded.

* feat: make --specific-settings and --skip-settings mutually exclusive

* fix: fix error reporting when HTTP retry failed

* fix: fix linter issue (colliding variable name)

* fix: fix setting of download options

* chore: remove duplication in download logic

* fix: added setting of missing skipSetting field in case of Download by manifest

* fix: chore fix failing unit test

* chore: added very simple e2e test to test --specific-api and --specific-settings flags

* chore: added additional test cases for downloading by manifest