feat(rest): Always retry failed Settings upsert
Add a very simple (dumb) solution to timing issue handling for settings.
Just like the Config API it can happen that a settings' dependency is not
found when we first try to create it - sometimes creation of things like
appliactions, management-zones, etc. takes a bit of time.

Other than Config client's implementation to try with different retrySettings
on different known timing issues, this change simply always retries Settings
creation and updates with the normal retry settings.

A better version of this shall be implementated later, but for now this
resolves the timing issues our own E2E tests encounter when creating SLOs as
Settings.