feat(manifest): Rename version and environment group fields in manifest files
For consistency and readability several fields in manifest files are renamed.
snake case is replaced with camel case to be consistent within all configs and similar to the
formatting of most k8s CRDs which we assume users are familiar with.

For clarity environment groups are renamed, making it clear by toplevel name that it is a list
of env groups and using more consistent and clear naming within the group definition.